Extract WeatherContainer component
