+++ src/features/weather/WeatherNow.jsx
@@ -1,16 +1,8 @@
-import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useParams } from "react-router-dom";
-import { getWeatherById } from "./weatherSlice";
+import React from "react";
+import { useSelector } from "react-redux";
 
 const WeatherNow = () => {
   const weatherNow = useSelector((state) => state.weather.now);
-  const dispatch = useDispatch();
-  const { id } = useParams();
-
-  useEffect(() => {
-    dispatch(getWeatherById(id));
-  }, [dispatch, id]);
 
   if (!weatherNow) {
     return <div>Loading...</div>;
